refactor: Ensure that only main window is tied to rootViewState
